(PC-32988) fix(offerToHeadlineOfferData): fix naming and upgrade coverage
