--- src/features/headlineOffer/adapters/headlineOfferData.native.test.ts
@@ -1,30 +1,0 @@
-import { offerToHeadlineOfferData } from 'features/headlineOffer/adapters/headlineOfferData'
-import { mockLabelMapping, mockMapping } from 'features/headlineOffer/fixtures/mockMapping'
-import { Currency } from 'shared/currency/useGetCurrencyToDisplay'
-import { offersFixture } from 'shared/offer/offer.fixture'
-
-describe('headlineOfferData', () => {
-  it('should transform headline offer data correctly', () => {
-    const result = offerToHeadlineOfferData({
-      offer: offersFixture[0],
-      transformParameters: {
-        mapping: mockMapping,
-        labelMapping: mockLabelMapping,
-        currency: Currency.EURO,
-        euroToPacificFrancRate: 10,
-        userLocation: { latitude: 1, longitude: 1 },
-      },
-    })
-
-    expect(result).toEqual({
-      id: '102280',
-      categoryId: 'LIVRE',
-      category: 'Livre papier',
-      imageUrl:
-        'https://storage.googleapis.com/passculture-metier-prod-production-assets-fine-grained/thumbs/mediations/CDNQ',
-      offerTitle: 'La nuit des temps',
-      price: 'Dès 28\u00a0€',
-      distance: '900+ km',
-    })
-  })
-})
